"catalog of the XML schema of Ofbiz for Eclipse" https://issues.apache.org/jira/browse/OFBIZ-4760 by Francis ANDRE
This is the catalog of the XML schema of OFBiz for Eclipse. Just import it and this will avoid Eclipse to go to the web for checking the xml files against their schema.

jleroux: I decided to commit, because it's a very little file and add some interesting help to Eclipse devs. For instance when disconnected and also makes sure you use the schema version corresponding to checked out code.
I added a comment about its use in the file itself and a link to wiki

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1326300 13f79535-47bb-0310-9956-ffa450edef68