New feature: Convert an XSD file to Java source code. The feature creates immutable Java object models that represent the XML file defined by the schema. It supports an optional Visitor implementation.
The feature can be accessed in Web Tools by clicking on the XSD to Java link.

The Java source code is displayed in a text area - where it can be copied-and-pasted to your IDE. The source code is fairly complete, but you will have to fill in some missing bits.

If anyone wants to adapt the feature to make it a command-line utility, then that would be welcome.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1489362 13f79535-47bb-0310-9956-ffa450edef68