Well, this is the start of converting ofbiz to use maven for building and packaging.  There's a workable parent-pom, and I have converted framework/start to maven as well.  To use this, the following commands are useful:
* mvn clean
* mvn compile
* mvn package

All 3 can be run in either ${ofbiz.home.dir}, or in framework/start.
The trick that made this simple, was that I was able to redefine maven's
default file layout away from src/main/java and target/*, to something
more ofbiz specific.  This way of doing things is just to get things
going; eventually, we should adopt a more standard maven like layout.

Another thing not dealt with in these first 2 files, is downloading of
dependencies.  The top-level and framework/start have no external
requirements, so it allowed me to get going quite quickly.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1674216 13f79535-47bb-0310-9956-ffa450edef68