Applied patch for OFBIZ-6662: PartyRelationship.securityGroupId not added when adding a PartyRelationship.
When adding a PartyRelationship the value selected for SecurityGroupId is ignored as createPartyRelationship service expects the parameter to be named securityGroupId. The AddOtherPartyRelationship form incorrectly names the field groupId causing the value to be missed in the set-nonpk-fields call in applications/party/script/org/ofbiz/party/party/PartyServices.xml.

Thanks Forrest Rae for spotting this and providing the patch.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1706694 13f79535-47bb-0310-9956-ffa450edef68