A patch from Christian Carlow for "Paginator.getListLimits listSize gets set to viewSize for FindServices.performFindList" https://issues.apache.org/jira/browse/OFBIZ-6422
This patch adds code to Paginator to account for the entryList being instanceOf List and extracts from the context.results object the original listSize.

jleroux, important note from OFBIZ-6423 which depends on OFBIZ-6422: it is necessary to assign a value (true makes sense) to override-list-size attribute so that FormRenderer.renderItemRows sets the lowIndex correctly, because once the results of performFindList are displayed, otherwise pages > 0 are rendered as empty.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1710413 13f79535-47bb-0310-9956-ffa450edef68