A slightly modified patch from Pierre Smits for <<comments on ordering products should be store driven>> https://issues.apache.org/jira/browse/OFBIZ-6924
Currently, in ecommerce/tiny-gismo-GZ-1000-p a comment field is shown.
This is a result of the following code in productdetail.ftl
{code}
<#assign commentEnable = Static["org.ofbiz.entity.util.EntityUtilProperties"].getPropertyValue("order", "order.item.comment.enable", delegator)>
            <#if commentEnable.equals("Y")>
                <#assign orderItemAttr = Static["org.ofbiz.entity.util.EntityUtilProperties"].getPropertyValue("order", "order.item.attr.prefix", delegator)>
                ${uiLabelMap.CommonComment}&nbsp;<input type="text" name="${orderItemAttr}comment"/>
            </#if>
{code}
Based on the code excerpt this is set by a property value in the order component. However this should be based on a configuration setting in the store.

jleroux: I simply changed the place of the ProductStore.allowComment field (not after the old deprecated fields) and used a description instead of an XML comment there.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1735435 13f79535-47bb-0310-9956-ffa450edef68