Fix the gradle build script on multiple reported issues - OFBIZ-7534
We were lucky to discover some problems in the build script on the
windows operating system (Thank you Jacques Le Roux). After investigating
the matter we realized that among other problems the process name (executable)
is massive and leads to problems on windows. However, this is not the only
problem with the current build script setup. Sometimes exeucting tasks in a combo
fail due to a dependency-tree problem with the current implementation.

The solution was to create a Jar file that holds the classpath and hides it from
the process name. Hence, instead of a massive classpath shown in the command process
it is all declared inside a single jar file. This lead to multiple improvements in the
build script including:

- The process name is now much shorter and cleaner both on Windows and Unix-like systems
- The task teminateOfbiz now has a much cleaner implementation on Unix systems
- Executing any tasks in combos would still maintain the correct ordering of executing graph

The most tricky part to figure out in the implementation is how to build the Jar Class-Path
for each operating system. Unfortunately, on Windows it was extremely tricky as windows
accepts weird naming conventions on the Class-Path manifest input. For example, \C:\whatever
will work but C:\whatever will not work

So the main difference in this implementation now is that the ofbiz server commands will
retrieve the classpath from a jar, and not directly construct it in the command line


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1752125 13f79535-47bb-0310-9956-ffa450edef68