This removes the pos component.
Most is done already. Test pass but testFopMacroLibrary but I don't see how it could be related to these changes, could be a quirk on Windows, BuilBot will tell us. Anyway, I will later double-check to be sure, also to see if I did not miss anything else.

I have voluntarily left 2 or 3 moot points
I was unable to remove possible POS references in <property key="FieldDescription.ProductStore.isImmediatelyFulfilled"> for certain language I'm unable to read.
I have to check if we should keep or not this code snippet
                <!-- if inventory info is not available for all of the products, then don't do the issuance, ie there has to be at least SOME inventory info in the database to facilitate inventory-less POS terminals -->
                <entity-count entity-name="InventoryItem" count-field="iiCount">
                    <condition-expr field-name="facilityId" operator="equals" from-field="orderHeader.originFacilityId"/>
                </entity-count>
I simply commented out twice this snippet in InvoiceServices.java, because I wonder if it would not be useful when the webPos will be able to issue money change
//                if (null == adj.get("amount")) { // JLR 17/4/7 : fix a bug coming from POS in case of use of a discount (on item(s) or sale, sale here) and a cash amount higher than total (hence issuing change)
//                    continue;
//                }



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1754402 13f79535-47bb-0310-9956-ffa450edef68