Improved: local library classpath setup in OFBiz
Now that all local jars are removed from OFBiz, there is no need to either
keep the "lib" directory in components or to automatically load these libs
if they exist. If users want to add custom libraries, then can place them
directly in the top level directory "lib".

Hence this commit removes the remaining lib directories in the accounting,
example and ldap components and fixes build.gradle to not iterate the lib
directories in components for loading jars inside of them.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1761302 13f79535-47bb-0310-9956-ffa450edef68